+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## [Unreleased]
+
+### Added
+
+### Changed
+
+### Removed
+
+### Fixed
+
+
+
+
+
+
+
+
 ## [v2.4.2] 2025-09-01
 
 ### Changed