+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 ### Fixed
 
 * Cell tracking module: fix option to select missing vertices in the graph (when showing results in napari).
-
+* About VLabApp: fix link to documentation on Windows.